integer math: fix handling of inf as integer
fix handling of conversion of int to intergers


git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@9842 a380766d-ff14-0410-b294-a243070f3f08

+++ common/Source/lang/LangPrimSource/PyrMathPrim.cpp
@@ -590,7 +590,7 @@ int mathClipInt(struct VMGlobals *g, int numArgsPushed)
 		if (err) return err;
 		err = slotDoubleVal(c, &hi);
 		if (err) return err;
-		SetFloat(a, sc_clip(slotRawInt(a), lo, hi));
+		SetFloat(a, sc_clip((double)slotRawInt(a), lo, hi));
 	}
 	return errNone;
 }
